requirements.txt
fastapi
uvicorn
speechmatics-rt
httpx
python-dotenv
requests
google-genai
